Ajout de l'icone pour les relations

+++ 2.Application/plurinotes/models/p_core.cpp
@@ -78,7 +78,7 @@ Version* PluriNotes::getNoteVersionByID(QString id, QString numV) {
 
     for(unsigned int i = 0; i < n->getVersions()->size() ; i++ ) {
         Version *v = n->getVersions()->at(i);
-        if(v->getNumVersion() == numV.toInt()) return v;
+        if(v->getNumVersion() == (unsigned)numV.toInt()) return v;
     }
 
     return 0;
